# artifacts/api-server/src/routes/programs.ts
import { Router, type IRouter } from "express";
import { db } from "../lib/db";
import { CreateProgramBody } from "@workspace/api-zod";

const router: IRouter = Router();

function escapeHtml(s: string): string {
  return s
    .replace(/&/g, "&amp;")
    .replace(/</g, "&lt;")
    .replace(/>/g, "&gt;")
    .replace(/"/g, "&quot;");
}

function buildProgramHtml(p: {
  churchName: string;
  sermonTitle: string;
  preacher: string;
  serviceDate: string;
  scripture1?: string | null;
  scripture2?: string | null;
  scripture3?: string | null;
  mainPoints?: string | null;
  announcements?: string | null;
  offeringTheme?: string | null;
}): string {
  const scriptures = [p.scripture1, p.scripture2, p.scripture3]
    .filter(Boolean)
    .map((s) => `<li>${escapeHtml(s as string)}</li>`)
    .join("");
  const points = (p.mainPoints || "")
    .split(/\r?\n/)
    .map((s) => s.trim())
    .filter(Boolean)
    .map((s, i) => `<li><strong>${i + 1}.</strong> ${escapeHtml(s)}</li>`)
    .join("");
  const announcements = (p.announcements || "")
    .split(/\r?\n/)
    .map((s) => s.trim())
    .filter(Boolean)
    .map((s) => `<li>${escapeHtml(s)}</li>`)
    .join("");
  return `<!doctype html><html><head><meta charset="utf-8"/><title>${escapeHtml(p.sermonTitle)} — ${escapeHtml(p.churchName)}</title>
<style>
  body { font-family: Georgia, 'Times New Roman', serif; max-width: 720px; margin: 2rem auto; padding: 1.5rem; color: #0F172A; line-height: 1.6; }
  h1 { font-size: 28px; margin: 0 0 0.25rem; color: #0F172A; }
  h2 { font-size: 14px; letter-spacing: 0.15em; text-transform: uppercase; color: #F59E0B; margin: 2rem 0 0.5rem; border-bottom: 2px solid #F59E0B; padding-bottom: 0.25rem; }
  .meta { color: #475569; font-size: 14px; margin-bottom: 0.5rem; }
  .preacher { font-style: italic; color: #475569; }
  ul { padding-left: 1.25rem; }
  .footer { margin-top: 2.5rem; text-align: center; font-size: 12px; color: #94A3B8; }
</style></head><body>
<header>
  <h1>${escapeHtml(p.churchName)}</h1>
  <div class="meta">Sunday Bulletin · ${escapeHtml(p.serviceDate)}</div>
</header>
<h2>Today's Message</h2>
<p><strong>${escapeHtml(p.sermonTitle)}</strong></p>
<p class="preacher">Preached by ${escapeHtml(p.preacher)}</p>
${scriptures ? `<h2>Scripture Readings</h2><ul>${scriptures}</ul>` : ""}
${points ? `<h2>Main Points</h2><ul>${points}</ul>` : ""}
${announcements ? `<h2>Announcements</h2><ul>${announcements}</ul>` : ""}
${p.offeringTheme ? `<h2>Offering Theme</h2><p>${escapeHtml(p.offeringTheme)}</p>` : ""}
<div class="footer">Generated by ChurchOS · ${escapeHtml(p.churchName)}</div>
</body></html>`;
}

router.get("/programs", (req, res) => {
  const church = (req.query.church as string) || "Demo Church Lagos";
  const rows = db
    .prepare("SELECT * FROM programs WHERE churchName = ? ORDER BY serviceDate DESC, id DESC")
    .all(church);
  res.json(rows);
});

router.get("/programs/:id", (req, res) => {
  const row = db.prepare("SELECT * FROM programs WHERE id = ?").get(Number(req.params.id));
  if (!row) { res.status(404).json({ error: "Not found" }); return; }
  res.json(row);
});

router.post("/programs", (req, res) => {
  const body = CreateProgramBody.parse(req.body);
  const html = buildProgramHtml(body);
  const info = db
    .prepare(
      `INSERT INTO programs (churchName, sermonTitle, preacher, serviceDate, scripture1, scripture2, scripture3, mainPoints, announcements, offeringTheme, generatedHtml)
       VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)`,
    )
    .run(
      body.churchName,
      body.sermonTitle,
      body.preacher,
      body.serviceDate,
      body.scripture1 ?? null,
      body.scripture2 ?? null,
      body.scripture3 ?? null,
      body.mainPoints ?? null,
      body.announcements ?? null,
      body.offeringTheme ?? null,
      html,
    );
  const row = db.prepare("SELECT * FROM programs WHERE id = ?").get(info.lastInsertRowid);
  res.status(201).json(row);
});

export default router;