# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ChurchOS API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Visitor {
  id: number;
  churchName: string;
  fullName: string;
  phone: string;
  email?: string | null;
  howHeard?: string | null;
  firstTime: boolean;
  visitDate: string;
  status: string;
  cellLeaderEmail?: string | null;
  notes?: string | null;
  createdAt: string;
}

export interface CreateVisitorBody {
  churchName: string;
  fullName: string;
  phone: string;
  email?: string;
  howHeard?: string;
  firstTime?: boolean;
}

export interface VisitorsSummary {
  total: number;
  newCount: number;
  contactedCount: number;
  returnedCount: number;
  coldCount: number;
}

export interface Program {
  id: number;
  churchName: string;
  sermonTitle: string;
  preacher: string;
  serviceDate: string;
  scripture1?: string | null;
  scripture2?: string | null;
  scripture3?: string | null;
  mainPoints?: string | null;
  announcements?: string | null;
  offeringTheme?: string | null;
  generatedHtml: string;
  createdAt: string;
}

export interface CreateProgramBody {
  churchName: string;
  sermonTitle: string;
  preacher: string;
  serviceDate: string;
  scripture1?: string;
  scripture2?: string;
  scripture3?: string;
  mainPoints?: string;
  announcements?: string;
  offeringTheme?: string;
}

export interface Giving {
  id: number;
  churchName: string;
  donorName: string;
  donorPhone: string;
  amount: number;
  category: string;
  status: string;
  ref: string;
  createdAt: string;
}

export interface InitiateGivingBody {
  churchName: string;
  donorName: string;
  donorPhone: string;
  amount: number;
  category: string;
}

export interface GivingCategoryTotal {
  category: string;
  total: number;
}

export interface GivingSummary {
  totalThisMonth: number;
  totalAllTime: number;
  byCategory: GivingCategoryTotal[];
}

export interface Member {
  id: number;
  churchName: string;
  fullName: string;
  phone: string;
  email?: string | null;
  cellGroup: string;
  cellLeaderEmail?: string | null;
  joinDate?: string | null;
  lastAttendance?: string | null;
  consecutiveMisses: number;
  status: string;
  createdAt: string;
}

export interface CreateMemberBody {
  churchName: string;
  fullName: string;
  phone: string;
  email?: string;
  cellGroup: string;
  cellLeaderEmail?: string;
}

export interface MembersSummary {
  total: number;
  active: number;
  atRisk: number;
  ghost: number;
  inactive: number;
}

export interface AttendanceEntry {
  memberId: number;
  present: boolean;
}

export interface SaveAttendanceBody {
  serviceDate: string;
  entries: AttendanceEntry[];
}

export interface AttendanceResult {
  presentCount: number;
  absentCount: number;
  alertsSent: number;
}

export interface Sermon {
  id: number;
  churchName: string;
  title: string;
  preacher: string;
  sermonDate: string;
  scripture: string;
  seriesName?: string | null;
  description?: string | null;
  audioFilename?: string | null;
  playCount: number;
  createdAt: string;
}

export interface CreateSermonBody {
  churchName: string;
  title: string;
  preacher: string;
  sermonDate: string;
  scripture: string;
  seriesName?: string;
  description?: string;
  audioFilename?: string;
}

export interface Notification {
  id: number;
  type: string;
  recipient: string;
  subject: string;
  body: string;
  sentAt: string;
  read: boolean;
}

export type ListVisitorsParams = {
  church?: string;
};

export type UpdateVisitorStatusBody = {
  status: string;
};

export type GetVisitorsSummaryParams = {
  church?: string;
};

export type ListProgramsParams = {
  church?: string;
};

export type ListGivingParams = {
  church?: string;
};

export type GetGivingSummaryParams = {
  church?: string;
};

export type ListMembersParams = {
  church?: string;
  status?: string;
};

export type GetMembersSummaryParams = {
  church?: string;
};

export type SeedMembersBody = {
  church: string;
};

export type SeedMembers200 = {
  inserted: number;
};

export type ListSermonsParams = {
  church?: string;
};

export type IncrementSermonPlay200 = {
  playCount: number;
};

export type GetUnreadNotificationsCount200 = {
  unread: number;
};

export type JoinWaitlistBody = {
  churchName: string;
  adminEmail: string;
};

export type JoinWaitlist201 = {
  ok: boolean;
};